# openapi/esim_go_schema_v2_5.yaml
openapi: 3.0.1
info:
  title: eSIM-Go API
  version: 2.5.0
  description: |
    To access the eSIMGo API, you need to authenticate your requests using an API key.
    Here's a brief overview of the authentication process:
      1. Assuming that you have created account already:
      - Log into your eSIMGo account at https://sso.esim-go.com/login.
      - Navigate to Account Settings -> API Details to find your API key.
      - Keep your API key secure and avoid sharing it with others.
      2. API Key Usage:
      - Include your API key in the header of all eSIMGo API requests.
      - Use the header key 'X-API-KEY' with your API key as the value.
      3. Security Scheme:
      - The eSIMGo API uses an HTTP security scheme type for authentication.
      4. Authorization Types:
      - The API supports 'apiKeyAuth' authorization type.
servers:
  - url: https://api.esim-go.com/v2.5
    description: eSIMGo API
...
paths:
...
  /esims:
    get:
      summary: List eSIMs
      description: |
        This endpoint retrieves all eSIMs currently assigned to your organization. It provides a comprehensive view of your eSIM inventory with flexible options for data retrieval and management.

        **Key Features:**
          1. Pagination: Efficiently manage large datasets by specifying the page number and items per page.
          2. Sorting: Customize the order of results using the 'direction' and 'orderBy' parameters.
          3. Filtering: Refine your search with multiple filter options for precise data retrieval.
      tags:
        - eSIMs
      parameters:
        - in: query
          name: page
          schema:
            type: string
          description: Page of ESIMs to return
        - in: query
          name: perPage
          schema:
            type: integer
            enum:
              - 10
              - 25
              - 50
              - 100
          description: Number of ESIMs to return per page
        - in: query
          name: direction
          schema:
            type: string
            enum:
              - asc
              - desc
          description: Direction of ordering
        - in: query
          name: orderBy
          schema:
            type: string
            enum:
              - iccid
          description: Name of column to order by
        - in: query
          name: filterBy
          schema:
            type: string
            enum:
              - iccid, customerRef, lastAction, actionDate, assignedDate
          description: Name of column to filter by. eSIMs can be filtered by ICCID, Customer Reference, Last Action (Bundle Refund, Bundle Applied, Bundle Revoked, eSIM Updated, eSIM Refreshed, eSIM Utilisation Alert), Last Action Date and SIM Assignment Date.
        - in: query
          name: filter
          schema:
            type: string
          description: Value to filter by
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: A list of your eSIMs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ESIMs"
              example:
                esims:
                  - iccid: "8901234567890123456"
                    customerRef: CUST001
                    lastAction: Bundle Applied
                    actionDate: "2023-05-15T14:30:00Z"
                    physical: false
                    assignedDate: "2023-05-01T10:00:00Z"
                    state: active
                  - iccid: "8909876543210987654"
                    customerRef: CUST002
                    lastAction: Activated
                    actionDate: "2023-05-16T09:15:00Z"
                    physical: true
                    assignedDate: "2023-05-10T11:30:00Z"
                    state: suspended
        "400":
          description: Bad Request - returned when incorrect data given
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl 'https://api.esim-go.com/v2.5/esims' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');

            let config = {
              method: 'get',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/esims',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/esims"
              method := "GET"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/esims"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("GET", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Get, "https://api.esim-go.com/v2.5/esims");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.get("https://api.esim-go.com/v2.5/esims")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
    put:
      summary: Update eSIM Details
      description: |
        A feature that enables the association of a unique identifier with an eSIM. This functionality allows for the integration of relevant operational data, such as customer order IDs.
      tags:
        - eSIMs
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: iccid
          in: query
          schema:
            type: string
          description: (Required) ICCID of eSIM
          example: 8944123456789012000
        - name: customerRef
          in: query
          schema:
            type: string
          description: (Required) New Customer Reference
          example: Test
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSimDetails"
            example:
              iccid: "8944123456789012345"
              customerRef: ref-123
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusMessage"
              example:
                status: success
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid; the user does not have any available Bundles left or the ICCID is not accessible by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "503":
          description: Processing - Please come back later or use the *Retry-After* (seconds) header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl -X PUT 'https://api.esim-go.com/v2.5/esims' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json' \
            -d '{"$ref":"#/components/schemas/UpdateSimDetails"}'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');
            let data = JSON.stringify({
              "$ref": "#/components/schemas/UpdateSimDetails"
            });

            let config = {
              method: 'put',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/esims',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              },
              data : data
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "strings"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/esims"
              method := "PUT"

              payload := strings.NewReader(`{"$ref":"#/components/schemas/UpdateSimDetails"}`)

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, payload)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/esims"

            payload = json.dumps({
              "$ref": "#/components/schemas/UpdateSimDetails"
            })
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("PUT", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Put, "https://api.esim-go.com/v2.5/esims");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent("{\"$ref\":\"#/components/schemas/UpdateSimDetails\"}", null, "application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.put("https://api.esim-go.com/v2.5/esims")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .body("{\"$ref\":\"#/components/schemas/UpdateSimDetails\"}")
              .asString();
...
  /esims/assignments:
    get:
      summary: Get eSIM Install Details
      description: |
        # eSIM SMDP+ Details Retrieval Endpoint

        This API endpoint retrieves eSIM SMDP+ details based on provided Order or Apply References.

        ## Key features:

        1. Input: One or multiple reference numbers can be submitted via query parameters.
        2. Output: For each reference, the system returns:
           - ICCID (Integrated Circuit Card Identifier)
           - SMDP+ (Subscription Manager Data Preparation) address
           - Matching ID
        3. Output format options:
           - Default: text/csv
           - Alternative formats: application/json, application/zip
           - Format selection is controlled via the 'Accept' header in the request
        4. Special functionality: When requesting 'application/zip', the response is a ZIP file containing QR code images in PNG format.

        Deactivated eSIMs will return a 410 Gone error.
      tags:
        - eSIMs
      parameters:
        - name: reference
          in: query
          schema:
            type: string
          description: (Required) Order Reference or Apply Reference
        - name: additionalFields
          in: query
          schema:
            type: string
          example: installUrl
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/eSIMDetailsInstallResponse"
              example:
                iccid: "8944123456789012345"
                matchingId: A1B2-C3D4-E5F6-G7H8
                smdpAddress: http://smdp.example.com
                profileStatus: Released
                pin: "1234"
                puk: "12345678"
                firstInstalledDateTime: "2023-06-15T14:30:00.000Z"
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid; the user does not have any available Bundles left or the ICCID is not accessible by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "410":
          description: Gone - returned when the eSIM has been deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "503":
          description: Processing - Please come back later or use the *Retry-After* (seconds) header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl 'https://api.esim-go.com/v2.5/esims/assignments' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');

            let config = {
              method: 'get',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/esims/assignments',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/esims/assignments"
              method := "GET"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/esims/assignments"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("GET", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Get, "https://api.esim-go.com/v2.5/esims/assignments");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.get("https://api.esim-go.com/v2.5/esims/assignments")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      x-displayName: API Key in Header
      description: |
        Your API Key can passed to the eSIMGo API by setting a Header with the key `X-API-Key` in your HTTP call.

        e.g.

        ```
        curl -H 'X-API-Key: $API_KEY' https://api.esim-go.com/v2.3/esims/$ICCID/bundles
        ```
  schemas:
...
    Message:
      type: object
      properties:
        message:
          description: status of api functions which do not return data
          type: string
...
    ESIMs:
      type: object
      properties:
        esims:
          type: array
          items:
            type: object
            properties:
              iccid:
                description: ICCID of ESIM
                type: string
              customerRef:
                description: Reference of ESIM
                type: string
              msisdn:
                description: MSISDN of ESIM. This only available on eSIMs with a Domestic Bundle applied (i.e. Voice and/or SMS)
                type: string
              lastAction:
                description: Last action performed on the ESIM (e.g. Bundle Applied)
                type: string
              actionDate:
                description: The date of the Last Action performed on the ESIM
                type: string
              physical:
                description: Type of SIM
                type: boolean
              assignedDate:
                description: The date of ESIM's first assignment
                type: string
              state:
                description: The current state of the eSIM
                type: string
                enum:
                  - active
                  - suspended
                  - deactivated
...
    UpdateSimDetails:
      type: object
      properties:
        iccid:
          type: string
          description: ICCID of eSIM
        customerRef:
          type: string
          description: New Customer Reference
...
    eSIMDetailsInstallResponse:
      type: object
      properties:
        iccid:
          type: string
          description: The ICCID (Integrated Circuit Card Identifier) of the eSIM
          example: "8944123456789012345"
        matchingId:
          type: string
          description: The matching ID for the eSIM profile
          example: A1B2-C3D4-E5F6-G7H8
        smdpAddress:
          type: string
          description: The address of the Subscription Manager Data Preparation (SM-DP+) server
          example: http://smdp.example.com
        profileStatus:
          type: string
          description: The current status of the eSIM profile
          example: Released
        pin:
          type: string
          description: The PIN (Personal Identification Number) for the eSIM
          example: "1234"
        puk:
          type: string
          description: The PUK (PIN Unlock Key) for the eSIM
          example: "12345678"
        firstInstalledDateTime:
          type: string
          format: date-time
          description: The date and time when the profile was first installed
          example: "2023-06-15T14:30:00.000Z"
...
    StatusMessage:
      type: object
      properties:
        status:
          description: status message of api request
          type: string
...